moved resource debug to shared
