+++ ui/app/src/components/shared/ResourceDebug.tsx
