+++ ui/app/src/components/workspaces/UserResourceItem.tsx
@@ -4,7 +4,7 @@ import { ApiEndpoint } from '../../models/apiEndpoints';
 import { Workspace } from '../../models/workspace';
 import { useAuthApiCall, HttpMethod } from '../../useAuthApiCall';
 import { UserResource } from '../../models/userResource';
-import { ResourceDebug } from './ResourceDebug';
+import { ResourceDebug } from '../shared/ResourceDebug';
 
 // TODO:
 // - This 'page' might die in place of a card on the Workspace services page - leave it alone for now
